Rename {test-,}awscopyclient to {test-,}s3copyclient (#366)
It's a misnomer.  AWSRuntime actually makes sense since it's tied to AWS.  S3CopyTask should theoretically be able to run anywhere (and it is, in the unit tests).
