+++ dss/events/chunkedtask/aws.py
@@ -5,8 +5,7 @@ import uuid
 
 import boto3
 
-from . import awscopyclient
-from . import _awstest, awsconstants
+from . import _awstest, awsconstants, s3copyclient
 from ._awsimpl import AWSRuntime
 from .runner import Runner
 
@@ -15,8 +14,8 @@ from .runner import Runner
 def get_clients():
     return {
         _awstest.AWS_FAST_TEST_CLIENT_NAME: _awstest.AWSFastTestTask,
-        awscopyclient.AWS_S3_COPY_CLIENT_NAME: awscopyclient.S3CopyTask,
-        awscopyclient.AWS_S3_COPY_AND_WRITE_METADATA_CLIENT_NAME: awscopyclient.S3CopyWriteBundleTask,
+        s3copyclient.AWS_S3_COPY_CLIENT_NAME: s3copyclient.S3CopyTask,
+        s3copyclient.AWS_S3_COPY_AND_WRITE_METADATA_CLIENT_NAME: s3copyclient.S3CopyWriteBundleTask,
     }
 
 logger = logging.getLogger()
