+++ dss/events/chunkedtask/s3copyclient.py
